+++ src/tokio_postgres.rs
@@ -213,7 +213,7 @@ impl DbQuery for TokioPostgresPool {
                     .and_then(|name| name.as_str().and_then(|name| Some(name))),
             ) {
                 (Some(column), Some(sql_type)) => {
-                    columns.insert(column.to_string(), sql_type.to_string())
+                    columns.insert(column.to_string(), sql_type.to_lowercase().to_string())
                 }
                 _ => {
                     return Err(DbError::DataError(format!(
@@ -529,300 +529,9 @@ impl DbQuery for TokioPostgresPool {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::core::{JsonValue, StringRow};
     use crate::params;
     use pretty_assertions::assert_eq;
-    use rust_decimal::dec;
     use serde_json::json;
-
-    #[tokio::test]
-    async fn test_text_column_query() {
-        let pool = TokioPostgresPool::connect("postgresql:///rltbl_db")
-            .await
-            .unwrap();
-        pool.execute_batch(
-            "DROP TABLE IF EXISTS test_table_text CASCADE;\
-             CREATE TABLE test_table_text ( value TEXT )",
-        )
-        .await
-        .unwrap();
-        pool.execute("INSERT INTO test_table_text VALUES ($1)", &["foo"])
-            .await
-            .unwrap();
-        let select_sql = "SELECT value FROM test_table_text WHERE value = $1";
-        let value = pool
-            .query_value(select_sql, &["foo"])
-            .await
-            .unwrap()
-            .as_str()
-            .unwrap()
-            .to_string();
-        assert_eq!("foo", value);
-
-        let string = pool.query_string(select_sql, &["foo"]).await.unwrap();
-        assert_eq!("foo", string);
-
-        let strings = pool.query_strings(select_sql, &["foo"]).await.unwrap();
-        assert_eq!(vec!["foo".to_owned()], strings);
-
-        let row = pool.query_row(select_sql, &["foo"]).await.unwrap();
-        assert_eq!(json!(row), json!({"value":"foo"}));
-
-        let rows = pool.query(select_sql, &["foo"]).await.unwrap();
-        assert_eq!(json!(rows), json!([{"value":"foo"}]));
-
-        let string_row = pool.query_string_row(select_sql, &["foo"]).await.unwrap();
-        assert_eq!(
-            StringRow::from([("value".to_owned(), "foo".to_owned())]),
-            string_row
-        );
-
-        let string_rows = pool.query_string_rows(select_sql, &["foo"]).await.unwrap();
-        assert_eq!(
-            vec![StringRow::from([("value".to_owned(), "foo".to_owned())])],
-            string_rows
-        );
-
-        // Clean up:
-        pool.drop_table("test_table_text").await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_integer_column_query() {
-        let pool = TokioPostgresPool::connect("postgresql:///rltbl_db")
-            .await
-            .unwrap();
-
-        pool.execute_batch(&format!(
-            "DROP TABLE IF EXISTS test_table_int CASCADE;\
-             CREATE TABLE test_table_int ( value_2 INT2, value_4 INT4, value_8 INT8 )",
-        ))
-        .await
-        .unwrap();
-        pool.execute(
-            "INSERT INTO test_table_int VALUES ($1, $2, $3)",
-            params![1_i16, 1_i32, 1_i64],
-        )
-        .await
-        .unwrap();
-
-        for column in ["value_2", "value_4", "value_8"] {
-            let params = match column {
-                "value_2" => params![1_i16],
-                "value_4" => params![1_i32],
-                "value_8" => params![1_i64],
-                _ => unreachable!(),
-            };
-            let select_sql = format!("SELECT {column} FROM test_table_int WHERE {column} = $1");
-            let value = pool
-                .query_value(&select_sql, params.clone())
-                .await
-                .unwrap()
-                .as_i64()
-                .unwrap();
-            assert_eq!(1, value);
-
-            let unsigned = pool.query_u64(&select_sql, params.clone()).await.unwrap();
-            assert_eq!(1, unsigned);
-
-            let signed = pool.query_i64(&select_sql, params.clone()).await.unwrap();
-            assert_eq!(1, signed);
-
-            let string = pool
-                .query_string(&select_sql, params.clone())
-                .await
-                .unwrap();
-            assert_eq!("1", string);
-
-            let strings = pool
-                .query_strings(&select_sql, params.clone())
-                .await
-                .unwrap();
-            assert_eq!(vec!["1".to_owned()], strings);
-
-            let row = pool.query_row(&select_sql, params.clone()).await.unwrap();
-            assert_eq!(json!(row), json!({column:1}));
-
-            let rows = pool.query(&select_sql, params.clone()).await.unwrap();
-            assert_eq!(json!(rows), json!([{column:1}]));
-        }
-
-        // Clean up:
-        pool.drop_table("test_table_int").await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_float_column_query() {
-        let pool = TokioPostgresPool::connect("postgresql:///rltbl_db")
-            .await
-            .unwrap();
-
-        // FLOAT8
-        pool.execute_batch(
-            "DROP TABLE IF EXISTS test_table_float CASCADE;\
-             CREATE TABLE test_table_float ( value FLOAT8 )",
-        )
-        .await
-        .unwrap();
-        pool.execute("INSERT INTO test_table_float VALUES ($1)", &[1.05_f64])
-            .await
-            .unwrap();
-        let select_sql = "SELECT value FROM test_table_float WHERE value > $1";
-        let value = pool
-            .query_value(select_sql, &[1.0_f64])
-            .await
-            .unwrap()
-            .as_f64()
-            .unwrap();
-        assert_eq!("1.05", format!("{value:.2}"));
-
-        let float = pool.query_f64(select_sql, &[1.0_f64]).await.unwrap();
-        assert_eq!(1.05, float);
-
-        let string = pool.query_string(select_sql, &[1.0_f64]).await.unwrap();
-        assert_eq!("1.05", string);
-
-        let strings = pool.query_strings(select_sql, &[1.0_f64]).await.unwrap();
-        assert_eq!(vec!["1.05".to_owned()], strings);
-
-        let row = pool.query_row(select_sql, &[1.0_f64]).await.unwrap();
-        assert_eq!(json!(row), json!({"value":1.05}));
-
-        let rows = pool.query(select_sql, &[1.0_f64]).await.unwrap();
-        assert_eq!(json!(rows), json!([{"value":1.05}]));
-
-        // FLOAT4
-        pool.execute_batch(
-            "DROP TABLE IF EXISTS test_table_float CASCADE;\
-             CREATE TABLE test_table_float ( value FLOAT4 )",
-        )
-        .await
-        .unwrap();
-        pool.execute("INSERT INTO test_table_float VALUES ($1)", &[1.05_f32])
-            .await
-            .unwrap();
-        let select_sql = "SELECT value FROM test_table_float WHERE value > $1";
-        let value = pool
-            .query_value(select_sql, &[1.0_f32])
-            .await
-            .unwrap()
-            .as_f64()
-            .unwrap();
-        assert_eq!("1.05", format!("{value:.2}"));
-
-        // Clean up:
-        pool.drop_table("test_table_float").await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_mixed_column_query() {
-        let pool = TokioPostgresPool::connect("postgresql:///rltbl_db")
-            .await
-            .unwrap();
-        pool.execute_batch(
-            "DROP TABLE IF EXISTS test_table_mixed CASCADE;\
-             CREATE TABLE test_table_mixed (\
-                 text_value TEXT,\
-                 alt_text_value TEXT,\
-                 float_value FLOAT8,\
-                 alt_float_value FLOAT8,\
-                 int_value INT8,\
-                 alt_int_value INT8,\
-                 bool_value BOOL,\
-                 alt_bool_value BOOL,\
-                 numeric_value NUMERIC,\
-                 alt_numeric_value NUMERIC\
-             )",
-        )
-        .await
-        .unwrap();
-        pool.execute(
-            r#"INSERT INTO test_table_mixed
-               (
-                 text_value,
-                 alt_text_value,
-                 float_value,
-                 alt_float_value,
-                 int_value,
-                 alt_int_value,
-                 bool_value,
-                 alt_bool_value,
-                 numeric_value,
-                 alt_numeric_value
-               )
-               VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)"#,
-            params!["foo", (), 1.05_f64, (), 1_i64, (), true, (), dec!(1.0), ()],
-        )
-        .await
-        .unwrap();
-
-        let select_sql = "SELECT text_value FROM test_table_mixed WHERE text_value = $1";
-        let value = pool
-            .query_value(select_sql, ["foo"])
-            .await
-            .unwrap()
-            .as_str()
-            .unwrap()
-            .to_string();
-        assert_eq!("foo", value);
-
-        let select_sql = r#"SELECT
-                              text_value,
-                              alt_text_value,
-                              float_value,
-                              alt_float_value,
-                              int_value,
-                              alt_int_value,
-                              bool_value,
-                              alt_bool_value,
-                              numeric_value,
-                              alt_numeric_value
-                            FROM test_table_mixed
-                            WHERE text_value = $1
-                              AND alt_text_value IS NOT DISTINCT FROM $2
-                              AND float_value > $3
-                              AND int_value > $4
-                              AND bool_value = $5
-                              AND numeric_value > $6"#;
-        let params = params!["foo", (), 1.0_f64, 0_i64, true, dec!(0.999)];
-
-        let row = pool.query_row(select_sql, params.clone()).await.unwrap();
-        assert_eq!(
-            json!(row),
-            json!({
-                "text_value": "foo",
-                "alt_text_value": JsonValue::Null,
-                "float_value": 1.05,
-                "alt_float_value": JsonValue::Null,
-                "int_value": 1,
-                "alt_int_value": JsonValue::Null,
-                "bool_value": true,
-                "alt_bool_value": JsonValue::Null,
-                "numeric_value": 1.0,
-                "alt_numeric_value": JsonValue::Null,
-            })
-        );
-
-        let rows = pool.query(select_sql, params.clone()).await.unwrap();
-        assert_eq!(
-            json!(rows),
-            json!([{
-                "text_value": "foo",
-                "alt_text_value": JsonValue::Null,
-                "float_value": 1.05,
-                "alt_float_value": JsonValue::Null,
-                "int_value": 1,
-                "alt_int_value": JsonValue::Null,
-                "bool_value": true,
-                "alt_bool_value": JsonValue::Null,
-                "numeric_value": 1.0,
-                "alt_numeric_value": JsonValue::Null,
-            }])
-        );
-
-        // Clean up:
-        pool.drop_table("test_table_mixed").await.unwrap();
-    }
 
     #[tokio::test]
     async fn test_aliases_and_builtin_functions() {
@@ -899,782 +608,5 @@ mod tests {
 
         // Clean up.
         pool.drop_table("test_table_indirect").await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_insert() {
-        let pool = TokioPostgresPool::connect("postgresql:///rltbl_db")
-            .await
-            .unwrap();
-        pool.execute_batch(
-            "DROP TABLE IF EXISTS test_insert CASCADE;\
-             CREATE TABLE test_insert (\
-                 text_value TEXT,\
-                 alt_text_value TEXT,\
-                 float_value FLOAT8,\
-                 int_value INT8,\
-                 bool_value BOOL\
-             )",
-        )
-        .await
-        .unwrap();
-
-        // Insert rows:
-        pool.insert(
-            "test_insert",
-            &["text_value", "int_value", "bool_value"],
-            &[
-                &json!({"text_value": "TEXT"}).as_object().unwrap(),
-                &json!({"int_value": 1, "bool_value": true})
-                    .as_object()
-                    .unwrap(),
-            ],
-        )
-        .await
-        .unwrap();
-
-        // Validate the inserted data:
-        let rows = pool
-            .query(r#"SELECT * FROM test_insert"#, ())
-            .await
-            .unwrap();
-        assert_eq!(
-            json!(rows),
-            json!([{
-                "text_value": "TEXT",
-                "alt_text_value": JsonValue::Null,
-                "float_value": JsonValue::Null,
-                "int_value": JsonValue::Null,
-                "bool_value": JsonValue::Null,
-            },{
-                "text_value": JsonValue::Null,
-                "alt_text_value": JsonValue::Null,
-                "float_value": JsonValue::Null,
-                "int_value": 1,
-                "bool_value": true,
-            }])
-        );
-
-        // Clean up.
-        pool.drop_table("test_insert").await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_insert_returning() {
-        let pool = TokioPostgresPool::connect("postgresql:///rltbl_db")
-            .await
-            .unwrap();
-        pool.execute_batch(
-            "DROP TABLE IF EXISTS test_insert_returning CASCADE;\
-             CREATE TABLE test_insert_returning (\
-                 text_value TEXT,\
-                 alt_text_value TEXT,\
-                 float_value FLOAT8,\
-                 int_value INT8,\
-                 bool_value BOOL\
-             )",
-        )
-        .await
-        .unwrap();
-
-        // No filtering:
-        let rows = pool
-            .insert_returning(
-                "test_insert_returning",
-                &["text_value", "int_value", "bool_value"],
-                &[
-                    &json!({"text_value": "TEXT"}).as_object().unwrap(),
-                    &json!({"int_value": 1, "bool_value": true})
-                        .as_object()
-                        .unwrap(),
-                ],
-                &[],
-            )
-            .await
-            .unwrap();
-        assert_eq!(
-            json!(rows),
-            json!([{
-                "text_value": "TEXT",
-                "alt_text_value": JsonValue::Null,
-                "float_value": JsonValue::Null,
-                "int_value": JsonValue::Null,
-                "bool_value": JsonValue::Null,
-            },{
-                "text_value": JsonValue::Null,
-                "alt_text_value": JsonValue::Null,
-                "float_value": JsonValue::Null,
-                "int_value": 1,
-                "bool_value": true,
-            }])
-        );
-
-        // With filtering:
-        let rows = pool
-            .insert_returning(
-                "test_insert_returning",
-                &["text_value", "int_value", "bool_value"],
-                &[
-                    &json!({"text_value": "TEXT"}).as_object().unwrap(),
-                    &json!({"int_value": 1, "bool_value": true})
-                        .as_object()
-                        .unwrap(),
-                ],
-                &["int_value", "float_value"],
-            )
-            .await
-            .unwrap();
-        assert_eq!(
-            json!(rows),
-            json!([{
-                "float_value": JsonValue::Null,
-                "int_value": JsonValue::Null,
-            },{
-                "float_value": JsonValue::Null,
-                "int_value": 1,
-            }])
-        );
-
-        // Clean up.
-        pool.drop_table("test_insert_returning").await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn drop_table() {
-        let pool = TokioPostgresPool::connect("postgresql:///rltbl_db")
-            .await
-            .unwrap();
-        let table1 = "test_drop1";
-        let table2 = "test_drop2";
-        pool.execute_batch(&format!(
-            "DROP TABLE IF EXISTS {table1} CASCADE;\
-             DROP TABLE IF EXISTS {table2} CASCADE;\
-             CREATE TABLE {table1} (\
-                 foo TEXT PRIMARY KEY\
-             );\
-             CREATE TABLE {table2} (\
-                 foo TEXT REFERENCES {table1}(foo)\
-             );",
-        ))
-        .await
-        .unwrap();
-
-        let columns = pool.columns(table1).await.unwrap();
-        assert_eq!(
-            columns,
-            ColumnMap::from([("foo".to_owned(), "text".to_owned())])
-        );
-        pool.drop_table(table1).await.unwrap();
-
-        let columns = pool.columns(table1).await.unwrap();
-        assert_eq!(columns.is_empty(), true);
-
-        // Clean up.
-        pool.drop_table(table2).await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_input_params() {
-        let pool = TokioPostgresPool::connect("postgresql:///rltbl_db")
-            .await
-            .unwrap();
-        pool.execute("DROP TABLE IF EXISTS test_input_params CASCADE", ())
-            .await
-            .unwrap();
-        pool.execute(
-            "CREATE TABLE test_input_params (\
-               bar TEXT,\
-               car INT2,\
-               dar INT4,\
-               far INT8,\
-               gar FLOAT4,\
-               har FLOAT8,\
-               jar NUMERIC,\
-               kar BOOL
-             )",
-            (),
-        )
-        .await
-        .unwrap();
-        pool.execute("INSERT INTO test_input_params (bar) VALUES ($1)", &["one"])
-            .await
-            .unwrap();
-        pool.execute(
-            "INSERT INTO test_input_params (far) VALUES ($1)",
-            &[1 as i64],
-        )
-        .await
-        .unwrap();
-        pool.execute("INSERT INTO test_input_params (bar) VALUES ($1)", ["two"])
-            .await
-            .unwrap();
-        pool.execute(
-            "INSERT INTO test_input_params (far) VALUES ($1)",
-            [2 as i64],
-        )
-        .await
-        .unwrap();
-        pool.execute(
-            "INSERT INTO test_input_params (bar) VALUES ($1)",
-            vec!["three"],
-        )
-        .await
-        .unwrap();
-        pool.execute(
-            "INSERT INTO test_input_params (far) VALUES ($1)",
-            vec![3 as i64],
-        )
-        .await
-        .unwrap();
-        pool.execute(
-            "INSERT INTO test_input_params (gar) VALUES ($1)",
-            vec![3 as f32],
-        )
-        .await
-        .unwrap();
-        pool.execute(
-            "INSERT INTO test_input_params (har) VALUES ($1)",
-            vec![3 as f64],
-        )
-        .await
-        .unwrap();
-        pool.execute(
-            "INSERT INTO test_input_params (jar) VALUES ($1)",
-            vec![dec!(3)],
-        )
-        .await
-        .unwrap();
-        pool.execute(
-            "INSERT INTO test_input_params (kar) VALUES ($1)",
-            vec![true],
-        )
-        .await
-        .unwrap();
-        pool.execute(
-            "INSERT INTO test_input_params \
-             (bar, car, dar, far, gar, har, jar, kar) \
-             VALUES ($1, $2, $3, $4, $5 ,$6, $7, $8)",
-            params![
-                "four",
-                123_i16,
-                123_i32,
-                123_i64,
-                123_f32,
-                123_f64,
-                dec!(123),
-                true,
-            ],
-        )
-        .await
-        .unwrap();
-        // Two alternative ways of specifying a NULL parameter:
-        let row = pool
-            .query_row(
-                "SELECT COUNT(1) AS count FROM test_input_params \
-                 WHERE bar IS NOT DISTINCT FROM $1 AND far IS NOT DISTINCT FROM $2",
-                params![ParamValue::Null, ()],
-            )
-            .await
-            .unwrap();
-        assert_eq!(json!({"count": 4}), json!(row));
-
-        // Clean up:
-        pool.drop_table("test_input_params").await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_primary_keys() {
-        let pool = TokioPostgresPool::connect("postgresql:///rltbl_db")
-            .await
-            .unwrap();
-        pool.execute_batch(
-            "DROP TABLE IF EXISTS test_primary_keys1 CASCADE;\
-             DROP TABLE IF EXISTS test_primary_keys2 CASCADE;\
-             CREATE TABLE test_primary_keys1 (\
-                 foo TEXT PRIMARY KEY\
-             );\
-             CREATE TABLE test_primary_keys2 (\
-                 foo TEXT,\
-                 bar TEXT,\
-                 car TEXT,
-                 PRIMARY KEY (foo, bar)\
-             )",
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(
-            pool.primary_keys("test_primary_keys1").await.unwrap(),
-            ["foo"]
-        );
-        assert_eq!(
-            pool.primary_keys("test_primary_keys2").await.unwrap(),
-            ["foo", "bar"]
-        );
-    }
-
-    #[tokio::test]
-    async fn test_update() {
-        let pool = TokioPostgresPool::connect("postgresql:///rltbl_db")
-            .await
-            .unwrap();
-        pool.execute_batch(
-            "DROP TABLE IF EXISTS test_update CASCADE;\
-             CREATE TABLE test_update (\
-               foo BIGINT PRIMARY KEY,\
-               bar BIGINT,\
-               car BIGINT,\
-               dar BIGINT,\
-               ear BIGINT
-             )",
-        )
-        .await
-        .unwrap();
-
-        pool.insert(
-            "test_update",
-            &["foo"],
-            &[
-                &json!({"foo": 1}).as_object().unwrap(),
-                &json!({"foo": 2}).as_object().unwrap(),
-                &json!({"foo": 3}).as_object().unwrap(),
-            ],
-        )
-        .await
-        .unwrap();
-
-        pool.update(
-            "test_update",
-            &["foo", "bar", "car", "dar", "ear"],
-            &[
-                &json!({
-                    "foo": 1,
-                    "bar": 10,
-                    "car": 11,
-                    "dar": 12,
-                    "ear": 13,
-                })
-                .as_object()
-                .unwrap(),
-                &json!({
-                    "foo": 2,
-                    "bar": 14,
-                    "car": 15,
-                    "dar": 16,
-                    "ear": 17,
-                })
-                .as_object()
-                .unwrap(),
-                &json!({
-                    "foo": 3,
-                    "bar": 18,
-                    "car": 19,
-                    "dar": 20,
-                    "ear": 21,
-                })
-                .as_object()
-                .unwrap(),
-            ],
-        )
-        .await
-        .unwrap();
-
-        let rows = pool.query("SELECT * from test_update", ()).await.unwrap();
-        assert_eq!(
-            json!(rows),
-            json!([
-                {
-                    "foo": 1,
-                    "bar": 10,
-                    "car": 11,
-                    "dar": 12,
-                    "ear": 13,
-                },
-                {
-                    "foo": 2,
-                    "bar": 14,
-                    "car": 15,
-                    "dar": 16,
-                    "ear": 17,
-                },
-                {
-                    "foo": 3,
-                    "bar": 18,
-                    "car": 19,
-                    "dar": 20,
-                    "ear": 21,
-                },
-            ])
-        )
-    }
-
-    #[tokio::test]
-    async fn test_update_returning() {
-        let pool = TokioPostgresPool::connect("postgresql:///rltbl_db")
-            .await
-            .unwrap();
-        pool.execute_batch(
-            "DROP TABLE IF EXISTS test_update_returning CASCADE;\
-             CREATE TABLE test_update_returning (\
-               foo BIGINT,\
-               bar BIGINT,\
-               car BIGINT,\
-               dar BIGINT,\
-               ear BIGINT,\
-               PRIMARY KEY (foo, bar)\
-             )",
-        )
-        .await
-        .unwrap();
-
-        pool.insert(
-            "test_update_returning",
-            &["foo", "bar", "car", "dar", "ear"],
-            &[
-                &json!({"foo": 1, "bar": 1}).as_object().unwrap(),
-                &json!({"foo": 2, "bar": 2}).as_object().unwrap(),
-                &json!({"foo": 3, "bar": 3}).as_object().unwrap(),
-            ],
-        )
-        .await
-        .unwrap();
-
-        let check_returning_rows = |rows: &Vec<JsonRow>| {
-            assert!(rows.iter().all(|row| {
-                [
-                    json!({
-                        "car": 10,
-                        "dar": 11,
-                        "ear": 12,
-                    }),
-                    json!({
-                        "car": 13,
-                        "dar": 14,
-                        "ear": 15,
-                    }),
-                    json!({
-                        "car": 16,
-                        "dar": 17,
-                        "ear": 18,
-                    }),
-                ]
-                .contains(&json!(row))
-            }));
-        };
-
-        check_returning_rows(
-            &pool
-                .update_returning(
-                    "test_update_returning",
-                    &["foo", "bar", "car", "dar", "ear"],
-                    &[
-                        &json!({
-                            "foo": 1,
-                            "bar": 1,
-                            "car": 10,
-                            "dar": 11,
-                            "ear": 12,
-                        })
-                        .as_object()
-                        .unwrap(),
-                        &json!({
-                            "foo": 2,
-                            "bar": 2,
-                            "car": 13,
-                            "dar": 14,
-                            "ear": 15,
-                        })
-                        .as_object()
-                        .unwrap(),
-                        &json!({
-                            "foo": 3,
-                            "bar": 3,
-                            "car": 16,
-                            "dar": 17,
-                            "ear": 18,
-                        })
-                        .as_object()
-                        .unwrap(),
-                    ],
-                    &["car", "dar", "ear"],
-                )
-                .await
-                .unwrap(),
-        );
-
-        // This is the same update as the first one above, just with the columns of the input
-        // rows to the update, as well as the rows themselves, specified in a different order.
-        pool.execute("DELETE FROM test_update_returning", ())
-            .await
-            .unwrap();
-
-        pool.insert(
-            "test_update_returning",
-            &["foo", "bar"],
-            &[
-                &json!({"foo": 1, "bar": 1}).as_object().unwrap(),
-                &json!({"foo": 2, "bar": 2}).as_object().unwrap(),
-                &json!({"foo": 3, "bar": 3}).as_object().unwrap(),
-            ],
-        )
-        .await
-        .unwrap();
-
-        check_returning_rows(
-            &pool
-                .update_returning(
-                    "test_update_returning",
-                    &["foo", "bar", "car", "dar", "ear"],
-                    &[
-                        &json!({
-                            "ear": 15,
-                            "bar": 2,
-                            "car": 13,
-                            "dar": 14,
-                            "foo": 2,
-                        })
-                        .as_object()
-                        .unwrap(),
-                        &json!({
-                            "foo": 1,
-                            "car": 10,
-                            "bar": 1,
-                            "ear": 12,
-                            "dar": 11,
-                        })
-                        .as_object()
-                        .unwrap(),
-                        &json!({
-                            "car": 16,
-                            "dar": 17,
-                            "ear": 18,
-                            "bar": 3,
-                            "foo": 3,
-                        })
-                        .as_object()
-                        .unwrap(),
-                    ],
-                    &["car", "dar", "ear"],
-                )
-                .await
-                .unwrap(),
-        );
-
-        // Final sanity check on the values of all columns:
-        let rows = pool
-            .query("SELECT * from test_update_returning", ())
-            .await
-            .unwrap();
-        assert!(rows.iter().all(|row| {
-            [
-                json!({
-                    "foo": 1,
-                    "bar": 1,
-                    "car": 10,
-                    "dar": 11,
-                    "ear": 12,
-                }),
-                json!({
-                    "foo": 2,
-                    "bar": 2,
-                    "car": 13,
-                    "dar": 14,
-                    "ear": 15,
-                }),
-                json!({
-                    "foo": 3,
-                    "bar": 3,
-                    "car": 16,
-                    "dar": 17,
-                    "ear": 18,
-                }),
-            ]
-            .contains(&json!(row))
-        }));
-    }
-
-    #[tokio::test]
-    async fn test_upsert() {
-        let pool = TokioPostgresPool::connect("postgresql:///rltbl_db")
-            .await
-            .unwrap();
-        pool.execute_batch(
-            "DROP TABLE IF EXISTS test_upsert CASCADE;\
-             CREATE TABLE test_upsert (\
-               foo BIGINT PRIMARY KEY,\
-               bar BIGINT,\
-               car BIGINT,\
-               dar BIGINT,\
-               ear BIGINT
-             )",
-        )
-        .await
-        .unwrap();
-
-        pool.insert(
-            "test_upsert",
-            &["foo"],
-            &[
-                &json!({"foo": 1}).as_object().unwrap(),
-                &json!({"foo": 2}).as_object().unwrap(),
-                &json!({"foo": 3}).as_object().unwrap(),
-            ],
-        )
-        .await
-        .unwrap();
-
-        pool.upsert(
-            "test_upsert",
-            &["foo", "bar", "car", "dar", "ear"],
-            &[
-                &json!({
-                    "foo": 1,
-                    "bar": 10,
-                    "car": 11,
-                    "dar": 12,
-                    "ear": 13,
-                })
-                .as_object()
-                .unwrap(),
-                &json!({
-                    "foo": 2,
-                    "bar": 14,
-                    "car": 15,
-                    "dar": 16,
-                    "ear": 17,
-                })
-                .as_object()
-                .unwrap(),
-                &json!({
-                    "foo": 3,
-                    "bar": 18,
-                    "car": 19,
-                    "dar": 20,
-                    "ear": 21,
-                })
-                .as_object()
-                .unwrap(),
-            ],
-        )
-        .await
-        .unwrap();
-
-        let rows = pool.query("SELECT * from test_upsert", ()).await.unwrap();
-        assert_eq!(
-            json!(rows),
-            json!([
-                {
-                    "foo": 1,
-                    "bar": 10,
-                    "car": 11,
-                    "dar": 12,
-                    "ear": 13,
-                },
-                {
-                    "foo": 2,
-                    "bar": 14,
-                    "car": 15,
-                    "dar": 16,
-                    "ear": 17,
-                },
-                {
-                    "foo": 3,
-                    "bar": 18,
-                    "car": 19,
-                    "dar": 20,
-                    "ear": 21,
-                },
-            ])
-        )
-    }
-
-    #[tokio::test]
-    async fn test_upsert_returning() {
-        let pool = TokioPostgresPool::connect("postgresql:///rltbl_db")
-            .await
-            .unwrap();
-        pool.execute_batch(
-            "DROP TABLE IF EXISTS test_upsert_returning CASCADE;\
-             CREATE TABLE test_upsert_returning (\
-               foo BIGINT,\
-               bar BIGINT,\
-               car BIGINT,\
-               dar BIGINT,\
-               ear BIGINT,\
-               PRIMARY KEY (foo, bar)\
-             )",
-        )
-        .await
-        .unwrap();
-
-        pool.insert(
-            "test_upsert_returning",
-            &["foo", "bar", "car", "dar", "ear"],
-            &[
-                &json!({"foo": 1, "bar": 1}).as_object().unwrap(),
-                &json!({"foo": 2, "bar": 2}).as_object().unwrap(),
-                &json!({"foo": 3, "bar": 3}).as_object().unwrap(),
-            ],
-        )
-        .await
-        .unwrap();
-
-        let rows = pool
-            .upsert_returning(
-                "test_upsert_returning",
-                &["foo", "bar", "car", "dar", "ear"],
-                &[
-                    &json!({
-                        "foo": 1,
-                        "bar": 1,
-                        "car": 10,
-                        "dar": 11,
-                        "ear": 12,
-                    })
-                    .as_object()
-                    .unwrap(),
-                    &json!({
-                        "foo": 2,
-                        "bar": 2,
-                        "car": 13,
-                        "dar": 14,
-                        "ear": 15,
-                    })
-                    .as_object()
-                    .unwrap(),
-                    &json!({
-                        "foo": 3,
-                        "bar": 3,
-                        "car": 16,
-                        "dar": 17,
-                        "ear": 18,
-                    })
-                    .as_object()
-                    .unwrap(),
-                ],
-                &["car", "dar", "ear"],
-            )
-            .await
-            .unwrap();
-        assert!(rows.iter().all(|row| {
-            [
-                json!({
-                    "car": 10,
-                    "dar": 11,
-                    "ear": 12,
-                }),
-                json!({
-                    "car": 13,
-                    "dar": 14,
-                    "ear": 15,
-                }),
-                json!({
-                    "car": 16,
-                    "dar": 17,
-                    "ear": 18,
-                }),
-            ]
-            .contains(&json!(row))
-        }));
     }
 }
